Keeping the shared functions under a single module
